requirements.txt
# Google Generative AI SDK
google-generativeai>=0.3.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# WebSocket Support
websockets>=12.0
python-multipart>=0.0.6

# Environment Variables
python-dotenv>=1.0.0

# Async Support
aiofiles>=23.2.1

# Logging
coloredlogs>=15.0.1